refactor(categories): rename and reorganize category-related classes and states for improved clarity
