+++ app/src/main/java/com/giraffe/tudeeapp/MainActivity.kt
@@ -17,9 +17,9 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import com.giraffe.tudeeapp.design_system.component.NavBar
 import com.giraffe.tudeeapp.design_system.theme.TudeeTheme
-import com.giraffe.tudeeapp.presentation.categories.screens.CategoriesScreen
 import com.giraffe.tudeeapp.presentation.tasks_by_category.TasksByCategoryScreen
 import com.giraffe.tudeeapp.presentation.tasks_by_category.TasksByCategoryViewModel
+import com.giraffe.tudeeapp.presentation.categories.CategoriesScreen
 import com.giraffe.tudeeapp.presentation.home.HomeScreen
 import com.giraffe.tudeeapp.presentation.navigation.Screen
 import com.giraffe.tudeeapp.presentation.splash.onboard.OnboardingScreen
