+++ app/src/main/java/com/giraffe/tudeeapp/presentation/categories/CategoriesScreenActions.kt
@@ -1,8 +1,8 @@
-package com.giraffe.tudeeapp.presentation.categories.state
+package com.giraffe.tudeeapp.presentation.categories
 
 import com.giraffe.tudeeapp.domain.model.Category
 
-interface CategoriesAction {
+interface CategoriesScreenActions {
     fun selectCategory(categoryId: Long)
     fun setBottomSheetVisibility(isVisible: Boolean)
     fun addCategory(category: Category)
