+++ app/src/main/java/com/giraffe/tudeeapp/presentation/categories/CategoriesScreenEvents.kt
@@ -1,0 +1,5 @@
+package com.giraffe.tudeeapp.presentation.categories
+
+sealed class CategoriesScreenEvents {
+    data class NavigateToTasksByCategoryScreen(val categoryId: Long) : CategoriesScreenEvents()
+}
