--- app/src/main/java/com/giraffe/tudeeapp/presentation/categories/state/CategoriesUiEvent.kt
@@ -1,5 +1,0 @@
-package com.giraffe.tudeeapp.presentation.categories.state
-
-sealed class CategoriesUiEvent {
-    data class NavigateToTasksByCategoryScreen(val categoryId: Long) : CategoriesUiEvent()
-}
